requirements.txt
asgiref==3.6.0
Django==4.1.4
djangorestframework==3.14.0
factory_boy==3.2.1
Faker==15.3.4
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
